Add payment-success notification
