+++ static/email/payment-success-icon.png
